执行命令

+++ stock/shares/model/shares_kdj.py
@@ -7,9 +7,9 @@ from .shares_name import SharesName
 
 class SharesKdj(models.Model):
     code = models.ForeignKey(SharesName, on_delete=models.CASCADE)
-    k = models.IntegerField(default=0)
-    d = models.IntegerField(default=0)
-    j = models.IntegerField(default=0)
+    k = models.FloatField(default=0)
+    d = models.FloatField(default=0)
+    j = models.FloatField(default=0)
     cycle_type = models.IntegerField(default=0, help_text='1.9,3,3')
     date_as = models.DateField()
